% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habcorr.R
\name{habcorr}
\alias{habcorr}
\title{Evaluate strength of correlation between HAB measure and other factor}
\usage{
habcorr(hab.data, hab.name = NULL, var.data, var.name = NULL)
}
\arguments{
\item{hab.data}{A data frame with columns "Date" (must be of class Date) and a measure of HAB (proxy measure of HAB, for example: cell count, chl-a concentration, phycocyanin concentration, etc.)}

\item{hab.name}{Name of the column for the measure of HAB}

\item{var.data}{A data frame with columns "Date" (must be of class Date) and some other factor to evaluate}

\item{var.name}{Name of the column for the parameter or factor}
}
\value{
The correlation (using non-parametric Spearman method) between the HAB measure and another variable
}
\description{
Evaluate strength of correlation between HAB measure and other factor
}
